refactor(cmd): migrate CLI to pli command tree
- rebuild cmd entrypoints on the pli command tree (tui, run,
  sessions, mcp, config, update)
- split the run loop into runloop; extract config, sessions,
  tui, and exit-code handling
- add github.com/pulseaiclub/pli dependency
