+++ cmd/bootstrap.go
@@ -17,14 +17,6 @@ import (
 
 const bootstrapDownloadTimeout = 5 * time.Minute
 
-// phi run exit codes.
-const (
-	ExitOK        = 0 // loop finished without errors
-	ExitError     = 1 // runtime / LLM / session error
-	ExitMaxRounds = 2 // model exceeded --max-rounds
-	ExitUsage     = 3 // config or CLI usage error
-)
-
 // HeadlessGate builds the permission gate for non-interactive entrypoints.
 // An empty policy mode defaults to headless-strict so Ask decisions fold to
 // Deny (Ask≡Deny); dangerously_allow_all is honored exactly like the TUI.
