+++ cmd/config.go
@@ -13,7 +13,6 @@ import (
 	"net/url"
 	"os"
 	"os/exec"
-	"os/signal"
 	"runtime"
 	"sort"
 	"strings"
@@ -21,7 +20,6 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/pulseaiclub/phi/internal/project"
 	"github.com/pulseaiclub/phi/internal/util"
 )
 
@@ -90,53 +88,6 @@ const (
 	modelListRequestLimit = 15 * time.Second
 	modelListBodyLimit    = int64(4 << 20)
 )
-
-// configCmd starts a local web server (loopback only) that edits config.yaml
-// in the browser.
-func configCmd(args []string) int {
-	for _, a := range args {
-		if a == "-h" || a == "--help" {
-			fmt.Fprintln(
-				os.Stdout,
-				"usage: phi config\n\nOpen the HTML config editor (starts a local web server on 127.0.0.1).",
-			)
-			return ExitOK
-		}
-	}
-	proj := project.GetDefaultProject()
-
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt)
-	defer stop()
-
-	var lc net.ListenConfig
-	ln, err := lc.Listen(ctx, "tcp", "127.0.0.1:0")
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "phi config:", err)
-		return ExitError
-	}
-	addr := ln.Addr().(*net.TCPAddr)
-	pageURL := fmt.Sprintf("http://127.0.0.1:%d/", addr.Port)
-	fmt.Fprintf(os.Stderr, "phi config: %s\n  config: %s\n  Ctrl-C to stop\n", pageURL, proj.Global().ConfigFile())
-	openBrowser(ctx, pageURL)
-
-	srv := &http.Server{
-		Handler:           &configHandler{configPath: proj.Global().ConfigFile()},
-		ReadHeaderTimeout: 10 * time.Second,
-	}
-
-	errc := make(chan error, 1)
-	go func() { errc <- srv.Serve(ln) }()
-	select {
-	case err := <-errc:
-		if err != nil && !errors.Is(err, http.ErrServerClosed) {
-			fmt.Fprintln(os.Stderr, "phi config:", err)
-			return ExitError
-		}
-	case <-ctx.Done():
-		_ = srv.Close()
-	}
-	return ExitOK
-}
 
 // configHandler serves the embedded editor page and its /api/config endpoints.
 type configHandler struct {
